Web Application/server/food-classification-101/requirements.txt
fastapi
uvicorn
pillow
tensorflow
numpy
python-multipart